Build: Let the users override the symbol versioning variant.
There are cases when the users want to decide themselves whether they want to
have the generic (even on GNU/Linux) or the linux (even if we do not recommend
that) symbol versioning variant. The former might be needed to circumvent
compiler issues (i.e. the compiler does not support all features that are
required for the linux versioning), the latter might help in overriding the
assumptions made in the configure script.