Add build-aux/version.txt
This is a new *duplicated* location for the package version.
Previously version was only in src/liblzma/api/lzma/version.h
from which it was extracted using build-aux/version.sh or using
a regex in CMakeLists.txt.

With Meson, this isn't practical anymore. While Meson can get
the version from an external command, such commands would need
to be available on non-POSIX targets as well. Even Python cannot
be required as the plan is to ensure that the Meson-based build
is compatible with the C99-only muon as well.